requirements.txt
# UI
PySide6>=6.6,<7

# OCR & documents
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.24.0

# Gmail / Outlook (ek indirme)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
msal>=1.28.0
